requirements.txt
reportlab>=3.6.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=9.0.0
